+++ public/index.php
@@ -686,6 +686,7 @@ if (\MuseDockPanel\Controllers\SetupController::needsSetup()) {
 \MuseDockPanel\Router::get('/docs/default-backups', 'DocsController@defaultBackups');
 \MuseDockPanel\Router::get('/docs/security-operations', 'DocsController@securityOperations');
 \MuseDockPanel\Router::get('/docs/firewall-operations', 'DocsController@firewallOperations');
+\MuseDockPanel\Router::get('/docs/panel-tls-dns01', 'DocsController@panelTlsDns01');
 \MuseDockPanel\Router::get('/docs/sync-archivos-lsyncd', 'DocsController@syncArchivosLsyncd');
 \MuseDockPanel\Router::get('/docs/profile-mfa', 'DocsController@profileMfa');